security-service/src/main/java/com/Naveen/security_service/service/AuthService.java
package com.Naveen.security_service.service;


import com.Naveen.security_service.entity.UserCredential;
import com.Naveen.security_service.repository.UserCredentialRepository;
import com.Naveen.security_service.utils.Constant;
import com.Naveen.security_service.utils.JwtService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

@Service
public class AuthService {

    @Autowired
    private UserCredentialRepository repository;

    @Autowired
    private PasswordEncoder passwordEncoder;

    @Autowired
    private JwtService jwtService;

    public String saveUser(UserCredential userCredential){
        userCredential.setPassword(passwordEncoder.encode(userCredential.getPassword()));
        repository.save(userCredential);
        return Constant.USER_SAVED;
    }

    public String generateToken(String username){
        return jwtService.generateToken(username);
    }

    public void validateToken(String token){
        jwtService.validateToken(token);
    }
}
